Remove "activeDeadlineSeconds" setting on pre-/post- upgrade job.
Why:
Currently we set both backoffLimit and activeDeadlineSeconds on pre-/post-
upgrade job. A job's activeDeadlineSeconds takes precedence over
backoffLimit, thus pre-/post- upgrade job will not deploy additional pods
once it reaches the time limit even if the backoffLimit is not hit.

When this happens, the following condition is set on the job, but
we are not checking that condition in version upgrade logic, thus
leading to operator stucks in upgrade inprogress state.

  conditions:
  - lastProbeTime: "2020-04-01T09:48:21Z"
    lastTransitionTime: "2020-04-01T09:48:21Z"
    message: Job was active longer than specified deadline
    reason: DeadlineExceeded
    status: "True"
    type: Failed
  failed: 5
  startTime: "2020-04-01T09:36:36Z"

What:
this change is to remove deadline seconds, thus simplify the logic
to cap the retries solely based on counts. If we ever needed to
have both time and count limits, then adding checking for deadline
exceeded condition in upgrade logic is required.